WoodMind_AgentSystem/requirements.txt
streamlit
pandas
psycopg2-binary
networkx
pyvis
pymorphy3
matplotlib
seaborn
openpyxl
transformers
torch